requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Caching
requests
python-nmap
python-dotenv
